pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "bgboxmaker"
version = "0.0.1"
authors = [
    { name="Samuel Plum", email="splum@taloria.com" },
]
description = "Board game tuck-box generator"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
    "typer",
    "pyyaml",
    "pillow"
]

[project.scripts]
bgboxmaker = "bgboxmaker.main:main"

[project.urls]
Homepage = "https://github.com/thetalorian/bgboxmaker"

[tool.coverage.run]
omit = [
    "__main__.py"
]